requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
google-genai
